quotient_of_num.py
divides number with decimal point

+++ quotient_of_num.py
@@ -1,6 +1,6 @@
 #ask user for input
-num1 = int(input("Enter num1: "))
-num2 = int(input("Enter num2: "))
+num1 = float(input("Enter num1: "))
+num2 = float(input("Enter num2: "))
 
 #divide two inputs
 print(num1 / num2)
